# perf(rueidisprob): reduce allocations in bloom filter operations