@rueian 2025-01-14
Thanks @nesty92. It is a cool optimization!